+++ CHANGELOG.md
@@ -4,7 +4,6 @@ Mapbox welcomes participation and contributions from everyone.
 
 # main
 
-
 # 11.2.1 March 15, 2024
 ## Bug fixes 🐞
 * Apply config expression to atmosphere properties.
@@ -15,6 +14,9 @@ Mapbox welcomes participation and contributions from everyone.
 
 
 # 11.3.0-rc.1
+## Features ✨ and improvements 🏁
+* [compose] Add layerPosition support in `GenericStyle` composable function.
+
 ## Bug fixes 🐞
 * Return `ViewAnnotationOptions.Builder` when calling `ViewAnnotationOptions.Builder.annotationAnchor` extension function.
 * [compose] Fix the `IndexOutOfBoundsException` because of `RootNode` of `MapboxMap` node tree being shared across multiple maps.